made it compatible with subclass in Application.groovy once more to support loading in dev environment. SpringApplicationRunListener may require jar to execute. This approach supports both methods and they share common trait
