+++ gradle.properties
@@ -1,4 +1,4 @@
 grailsVersion=3.1.12
 gradleWrapperVersion=2.14.1
 
-pluginVersion = 1.1.3-BUILD-SNAPSHOT
+pluginVersion = 1.1.4-BUILD-SNAPSHOT
