+++ src/main/groovy/grails/plugin/externalconfig/ExternalConfig.groovy
@@ -1,0 +1,12 @@
+package grails.plugin.externalconfig
+import groovy.transform.CompileStatic
+import org.springframework.context.EnvironmentAware
+import org.springframework.core.env.Environment
+
+@CompileStatic
+trait ExternalConfig implements EnvironmentAware, ExternallyConfigurable {
+    @Override
+    void setEnvironment(Environment environment) {
+       configureEnvironment(environment)
+    }
+}
